+++ tools/Test-PusfumeSource.ps1
@@ -559,23 +559,20 @@ Test-Condition ((Test-Path (Join-Path $repoRoot "tools\splice_bundle_resource.py
     $nativeBuildText -match '\$ParentChildMaterial = \$true' -and `
     $nativeBuildText -match '\$NoDonorTextureShadow = \$true' -and `
     $nativeBuildText -match 'make_spliced_child\.py' -and `
-    $nativeBuildText -match '"--expect-size", "496"' -and `
+    $nativeBuildText -match '"--expect-size", "768"' -and `
     $nativeBuildText -match '"--expect-size", "96"' -and `
     $nativeBuildText -match '"--expect-size", "128"' -and `
-    $nativeBuildText -match '"--expect-parent", "BBBF9694DA11465F"' -and `
+    $nativeBuildText -match '"--expect-parent", "3D25339231384C80"' -and `
     $nativeBuildText -match '"--expect-parent", "D97596A091982F4B"' -and `
     $nativeBuildText -match '"--expect-parent", "F85B289742D5D69A"' -and `
     $nativeBuildText -match 'hash:F72D636600F7F598' -and `
-    $nativeBuildText -match 'hash:D18F69CCD2253779' -and `
-    $nativeBuildText -match 'ED67ABE0A2542484=C263ECB79A8DCEC0' -and `
-    $nativeBuildText -match '4B7F05AED3F40BDF=A4215592F6297E57' -and `
-    $nativeBuildText -match 'A706B01BC822A417=818C87B860407405' -and `
-    $nativeBuildText -match 'tint_skin=0,0,1' -and `
-    $nativeBuildText -match 'tint_fur=0,0,1' -and `
-    $nativeBuildText -match 'tint_color_power=1' -and `
-    $nativeBuildText -match 'texture_map_990e13c4=C263ECB79A8DCEC0' -and `
-    $nativeBuildText -match 'texture_map_6e114674=A4215592F6297E57' -and `
-    $nativeBuildText -match 'texture_map_8f6fa466=818C87B860407405' -and `
+    $nativeBuildText -match 'hash:90BDF3BAC6F81BA8' -and `
+    $nativeBuildText -match 'DD74D8319F514D96=C263ECB79A8DCEC0' -and `
+    $nativeBuildText -match 'E334A8CB6BCB5E6D=A4215592F6297E57' -and `
+    $nativeBuildText -match '45FFAEEF53695A86=818C87B860407405' -and `
+    $nativeBuildText -match 'texture_map_02af90f8=C263ECB79A8DCEC0' -and `
+    $nativeBuildText -match 'texture_map_27b67fd2=818C87B860407405' -and `
+    $nativeBuildText -match 'texture_map_8bf37d8e=A4215592F6297E57' -and `
     $nativeBuildText -match '86FFDEB90C40C597=E0C4E09D80AE735B' -and `
     $nativeBuildText -match '258E4E4AEA37B1B8=3B3F6545AF6782F5' -and `
     $nativeBuildText -match 'E04C4FD132004376=DF5A1D6679E28376' -and `